Fix: TS diasteriomer expansion w/ reagent matching
Previously, this was not possible because the expansion was filtered
before matching to the reagents. Now, we match to reagents first and
filter after.

The reactant and product graphs are passed directly to the expand_stereo
function, so there is no longer a need for a separate
`expand_stereo_to_match_reagents` function.