Fix: Random geometry embedding failures
The failures seem to be caused by RDKit's "basic knowledge" filter,
which fails for molecules like propadiene (C=C=C). When the first
attempt fails, we now re-try the embedding with basic knowledge turned
off. That seems to work.

Also, we are now passing a random seed so that these errors are
consistent and reproducible.